refactor: extract param and slot argv rendering into params.py
Moves 16 top-level symbols (10 functions/helpers + 2 model types + 4
constants) from server.py into a new leaf module, src/comfy_mcp/params.py,
byte-for-byte: turning a caller's params/overrides/slots dict-or-string
input into the argv tokens comfy generate / run-template / workflow
set-slot / workflow vary take — _generate_param_args,
_run_template_param_args, _validate_param_key, _validate_generate_model
(+ _GENERATE_RESERVED_TARGETS, _GENERATE_META_FLAGS), and the structured
slot machinery: SlotOverride, SlotVariants, _slot_address_arg,
_slot_value_json, _SlotModel, _as_slot_model, _slot_override_arg,
_slot_variants_arg, _reject_non_json_array_slot (+ _SLOT_VALUE_EXAMPLE,
_SLOT_QUOTED_EXAMPLE, _MAX_PRECHECKED_SLOT_BYTES). Five blocks moved
verbatim from pre-move server.py (3082-3108, 3521-3629, 4307-4350,
9792-9937, 9991-10124); verified byte-for-byte present in params.py via
a source-substring check against pre-move HEAD.

SlotOverride/SlotVariants are name-imported into server
(`from .params import SlotOverride, SlotVariants`) rather than reached
module-qualified — the ONE carve-out this repo's extractions use for
public types (mirrors ComfyCliError from the errors.py commit): they
ride @mcp.tool() signatures (set_workflow_slot's overrides,
vary_workflow's slots), and with `from __future__ import annotations`
a module-qualified `params.SlotOverride` in the signature text would
change the exported MCP schema's $defs.

Everything else reaches server.py module-qualified (params._x), with
one recurring collision this commit hits for the first time: `params`
is also the PUBLIC ARGUMENT NAME on three tools whose signature is part
of the MCP schema (partner_generate, emit_partner_workflow,
run_template) — unlike the argv/errors extractions' local-variable
shadows, this local could not simply be renamed without itself being a
schema-breaking change. Fixed by a scoped `from . import params as
_params` at the top of each of those three function bodies, reaching
the module through that alias instead; the bare `params` name stays
the one used everywhere else in server.py (including a genuinely
renamable local in generate_image, whose internal `params` dict became
`template_params` since it is not part of that tool's public schema).
Disclosed inline at each of the four sites.

SCHEMA GATE (mandatory for this commit): discovered the SDK's real
tool-listing API — `mcp.server.mcpserver.MCPServer.list_tools()`
(async, returns each tool's `input_schema`/`output_schema`) — dumped
every one of the 52 registered tools' schemas to JSON before the move
and again after. Diff: EMPTY. Verdict: PASS, no schema drift.

Tests: tests/test_emit_partner_workflow.py's ONE monkeypatch
(`setattr(server, "_validate_generate_model", ...)`) moves to
`setattr(params, ...)` — patching the module object works through the
scoped `_params` alias too, since both names bind the same module.
tests/test_partner_catalog.py and tests/test_workflow.py rename their
read-only references (`server._GENERATE_RESERVED_TARGETS` ->
`params._GENERATE_RESERVED_TARGETS`; 7x
`server._MAX_PRECHECKED_SLOT_BYTES` -> `params._MAX_PRECHECKED_SLOT_BYTES`).
TypeVar dropped from server.py's imports (its only user, _SlotModel,
moved).

AGENTS.md's module table gains a params.py row ("Eight leaf modules"
-> "Nine").

Full suite: 1889 passed, 3 deselected. ruff check and ruff format
--check both clean.
